finish first pass
