+++ fabric-content-registries-v0/src/testmod/java/net/fabric/test/FabricContentRegistriesTest.java
@@ -2,7 +2,9 @@ package net.fabric.test;
 
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.registry.FlammableBlockRegistry;
+import net.fabricmc.fabric.api.registry.FuelRegistry;
 import net.minecraft.block.Blocks;
+import net.minecraft.item.Items;
 
 public class FabricContentRegistriesTest implements ModInitializer {
 
@@ -10,5 +12,8 @@ public class FabricContentRegistriesTest implements ModInitializer {
 	public void onInitialize() {
 		FlammableBlockRegistry.getDefaultInstance().add(Blocks.STONE, 100, 100);
 		FlammableBlockRegistry.getDefaultInstance().remove(Blocks.OAK_PLANKS);
+
+		FuelRegistry.INSTANCE.add(Items.APPLE, 200);
+		FuelRegistry.INSTANCE.remove(Blocks.OAK_PLANKS);
 	}
 }
