+++ fabric-loot-tables-v1/src/testmod/resources/fabric.mod.json
@@ -1,0 +1,13 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-loot-tables-v1-test",
+  "version": "0.0.0",
+  "name": "fabric-loot-tables-v1-test",
+  "environment": "*",
+  "entrypoints": {
+    "main": [
+      "net.fabric.test.LootEntryTestMod",
+      "net.fabric.test.LootTableTestMod"
+    ]
+  }
+}
