+++ fabric-models-v0/build.gradle
@@ -3,4 +3,6 @@ version = getSubprojectVersion(project, "0.1.0")
 
 dependencies {
 	compile project(path: ':fabric-api-base', configuration: 'dev')
+
+	testmodCompile project(path: ':fabric-events-lifecycle-v0', configuration: 'dev')
 }
