--- fabric-testmods/resources/mod.json
@@ -1,15 +1,0 @@
-{
-  "id": "fabric_test",
-  "name": "Fabric API Tests",
-  "version": "0.1.1",
-  "side": "universal",
-  "description": "A series of test mods to check Fabric works.",
-  "license": "Apache-2.0",
-  "initializers": [
-    "net.fabricmc.fabric.colormapper.ColorProviderMod",
-    "net.fabricmc.fabric.events.ServerEventMod",
-    "net.fabricmc.fabric.containers.ContainerMod",
-    "net.fabricmc.fabric.containers.ContainerModClient",
-    "net.fabricmc.fabric.loot.LootEntryMod"
-  ]
-}
